bug fixed

+++ app/screens/customer/screens/viewCustomer/index.jsx
@@ -9,11 +9,17 @@ import { db } from '../../../../hooks/firebase';
 
 import { Feather } from '@expo/vector-icons';
 import color from '../../../../style/color';
+import Header from '../../../../components/Header'
+import { useSelector } from 'react-redux';
+
+import app from '../../../../style/app'
 
 
 const ViewCustomer = () => {
   const { viewCustomer } = useRoute().params
   const { navigate } = useNavigation()
+
+  const { theme } = useSelector(state => state.user)
 
   const [contact, setContact] = useState({ ...viewCustomer })
   const [invoices, setInvoices] = useState([])
@@ -57,67 +63,87 @@ const ViewCustomer = () => {
     return grandTotal?.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
   }
 
-  const handleArchive = async () => {
-    let customerId = viewCustomer?.customerId
-
-    const id = JSON.parse(await AsyncStorage.getItem('recido_user'))?.user?.uid
-
-    setArchiveLoading(true)
-
-    let customers = (await getDoc(doc(db, 'users', id, 'customers', customerId))).data()
-
-    await setDoc(doc(db, 'users', id, 'customerArchive', customerId),
+  const handleArchive = () => {
+    Alert.alert('Archive customer', 'Are you sure you want to move this customer to your archive?', [
       {
-        ...customers,
-        archivedAt: serverTimestamp()
+        text: 'Cancel'
+      },
+      {
+        text: 'Move to archive',
+        onPress: async () => {
+          let customerId = viewCustomer?.customerId
+
+          const id = JSON.parse(await AsyncStorage.getItem('recido_user'))?.user?.uid
+
+          setArchiveLoading(true)
+
+          let customers = (await getDoc(doc(db, 'users', id, 'customers', customerId))).data()
+
+          await setDoc(doc(db, 'users', id, 'customerArchive', customerId),
+            {
+              ...customers,
+              archivedAt: serverTimestamp()
+            }
+          )
+
+          await deleteDoc(doc(db, 'users', id, 'customers', customerId))
+
+          setArchiveLoading(false)
+
+          Alert.alert('Contact has been moved to your archive successfully 🎉🎉')
+        }
       }
-    )
-
-    await deleteDoc(doc(db, 'users', id, 'customers', customerId))
-
-    setArchiveLoading(false)
-
-    Alert.alert('Contact has been moved to your archive successfully 🎉🎉')
+    ])
   };
 
   const handleUpdate = async () => {
-    const id = JSON.parse(await AsyncStorage.getItem('recido_user'))?.user?.uid
-    let customerId = viewCustomer?.customerId
+    Alert.alert('Customer update', 'Some changes have been made to this customer.\nDo you want to save?', [
+      {
+        text: 'Cancel'
+      },
+      {
+        text: 'Update customer',
+        onPress: async () => {
+          const id = JSON.parse(await AsyncStorage.getItem('recido_user'))?.user?.uid
+          let customerId = viewCustomer?.customerId
 
-    setUpdateLoading(true)
+          setUpdateLoading(true)
 
-    await updateDoc(doc(db, 'users', id, 'customers', customerId), {
-      ...contact
-    })
+          await updateDoc(doc(db, 'users', id, 'customers', customerId), {
+            ...contact
+          })
 
-    setUpdateLoading(false)
+          setUpdateLoading(false)
 
-    Alert.alert('Contact has been updated successfully 🎉🎉')
+          Alert.alert('Contact has been updated successfully 🎉🎉')
+        }
+      }
+    ])
   }
 
   return (
-    <View style={styles.container}>
-      <Header />
+    <View style={{ ...styles.container, backgroundColor: theme ? color.dark : color.mainBackground }}>
+      <Header title='Customers' />
 
       <ScrollView style={styles.scrollView} showsVerticalScrollIndicator={false}>
-        <Text style={styles.headingText}>Invoices</Text>
+        <Text style={{ ...styles.headingText, color: theme ? color.white : color.dark }}>{`${invoices[0]?.invoiceContact?.name}'s`} invoices ({invoices?.length})</Text>
         {
           invoices?.map((item, index) => (
-            <Pressable key={index} onPress={() => navigate('Create', { viewInvoice: item })} style={{ ...styles.list, paddingTop: 10 }}>
+            <Pressable key={index} onPress={() => navigate('Create', { viewInvoice: item })} style={{ ...styles.list, paddingTop: 10, backgroundColor: theme ? color.black : color.white }}>
               <View style={styles.left}>
-                <Text style={styles.boldText}>{item?.invoiceContact?.name}</Text>
-                <Text>#{item?.invoiceId}</Text>
+                <Text style={{ ...styles.boldText, color: theme ? color.white : color.dark }}>{item?.invoiceContact?.name}</Text>
+                <Text style={{ color: theme ? color.white : color.dark }}>#{item?.invoiceId}</Text>
               </View>
               <View style={styles.right}>
-                <Text style={styles.boldText}>{new Date(item?.date).toDateString()}</Text>
-                <Text>{calculateTotal(item)}</Text>
+                <Text style={{ ...styles.boldText, color: theme ? color.white : color.dark }}>{new Date(item?.date).toDateString()}</Text>
+                <Text style={{ color: theme ? color.white : color.dark }}>{calculateTotal(item)}</Text>
               </View>
             </Pressable>
           ))
         }
 
         <View style={styles.headingView}>
-          <Text style={styles.headingText}>Contact</Text>
+          <Text style={{ ...styles.headingText, color: theme ? color.white : color.dark }}>Contact</Text>
           <View style={styles.actionsView}>
             {
               archiveLoading ?
@@ -136,84 +162,93 @@ const ViewCustomer = () => {
             }
           </View>
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Name</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Name</Text>
           <TextInput
             placeholder='Name'
+            placeholderTextColor={theme ? color.white : color.dark}
             value={contact?.name}
             onChangeText={name => setContact({ ...contact, name })}
-            style={styles.inputViewTextInput}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             editable={false}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Email</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Email</Text>
           <TextInput
             placeholder='Email'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.email}
             onChangeText={email => setContact({ ...contact, email })}
-          />
+            />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Mobile</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Mobile</Text>
           <TextInput
             placeholder='Mobile'
-            style={styles.inputViewTextInput}
-            value={contact?.phoneNumbers[0]?.digits}
-            onChangeText={digits => setContact({ ...contact, phoneNumbers: [{ ...contact.phoneNumbers[0], digits }] })}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
+            value={contact?.phoneNumbers != undefined ? contact?.phoneNumbers[0]?.digits : contact?.phone}
+            onChangeText={digits => setContact(contact?.phoneNumbers != undefined ? { ...contact, phoneNumbers: [{ ...contact.phoneNumbers[0], digits }] } : { ...contact, phone: digits })}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Phone</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Phone</Text>
           <TextInput
             placeholder='Phone'
-            style={styles.inputViewTextInput}
-            value={contact?.phoneNumbers[0]?.number}
-            onChangeText={number => setContact({ ...contact, phoneNumbers: [{ ...contact.phoneNumbers[0], number }] })}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
+            value={contact?.phoneNumbers != undefined ? contact?.phoneNumbers[0]?.number : contact?.number}
+            onChangeText={number => setContact(contact?.phoneNumbers != undefined ? { ...contact, phoneNumbers: [{ ...contact.phoneNumbers[0], number }] } : { ...contact, number })}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Fax</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Fax</Text>
           <TextInput
             placeholder='Fax'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.fax}
             onChangeText={fax => setContact({ ...contact, fax })}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Country</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Country</Text>
           <TextInput
             placeholder='Country'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.country}
             onChangeText={country => setContact({ ...contact, country })}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>State</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>State</Text>
           <TextInput
             placeholder='State'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.country}
             onChangeText={state => setContact({ ...contact, state })}
-          />
+            />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>City</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>City</Text>
           <TextInput
             placeholder='City'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.country}
             onChangeText={city => setContact({ ...contact, city })}
           />
         </View>
-        <View style={styles.inputView}>
-          <Text style={styles.inputViewText}>Zip</Text>
+        <View style={app.inputView}>
+          <Text style={{ ...app.inputText, color: theme ? color.white : color.dark }}>Zip</Text>
           <TextInput
             placeholder='Zip'
-            style={styles.inputViewTextInput}
+            placeholderTextColor={theme ? color.white : color.dark}
+            style={{ ...app.input, color: theme ? color.white : color.dark }}
             value={contact?.country}
             onChangeText={zip => setContact({ ...contact, zip })}
           />
